Improve performance some, and remove dependency on floorf

+++ src/core/gpu_sw_backend.h
@@ -143,8 +143,8 @@ protected:
   // Rasterization
   //////////////////////////////////////////////////////////////////////////
   template<bool texture_enable, bool raw_texture_enable, bool transparency_enable, bool dithering_enable>
-  void ShadePixel(const GPUBackendDrawCommand* cmd, s32 x, s32 y, u8 color_r, u8 color_g, u8 color_b, float texcoord_x,
-                  float texcoord_y);
+  void ShadePixel(const GPUBackendDrawCommand* cmd, s32 x, s32 y, u8 color_r, u8 color_g, u8 color_b, u8 texcoord_x,
+                  u8 texcoord_y);
 
   template<bool texture_enable, bool raw_texture_enable, bool transparency_enable>
   void DrawRectangle(const GPUBackendDrawRectangleCommand* cmd);
